mpi/pmix: Add initial version of UCX support

+++ src/plugins/mpi/pmix/pmixp_dconn_tcp.h
@@ -42,7 +42,7 @@
 #include "pmixp_dconn.h"
 
 int pmixp_dconn_tcp_prepare(pmixp_dconn_handlers_t *handlers,
-			    char **ep_data, uint16_t *ep_len);
+			    char **ep_data, size_t *ep_len);
 void pmixp_dconn_tcp_finalize();
 
 #endif
